@date : 2026-01-26 @author : 김우렬 @version : 1.0.14:26012602 @comment : 주문전표 응답 데이터 타입 수정

+++ app/src/main/java/kr/co/kimberly/wma/network/model/CustomerModel.kt
@@ -7,7 +7,7 @@ data class CustomerModel (
     val customerCd: String, // 거래처 코드
     val custNm: String, // 거래처 명
     val customerNm: String, // 거래처 코드
-    val remainAmt: Int? = null, // 채권잔액
+    val remainAmt: Long? = null, // 채권잔액
     val slipNo: String? = null, // 전표 번호
     val totalAmount: Int? = null // 금액
 ): Serializable
